fix: Import header files

+++ src/main.cpp
@@ -1,9 +1,9 @@
 #include <cassert>
-#include "mat/mat3x3.cpp"
-#include "mat/mat4x4.cpp"
-#include "vec/vec2.cpp"
-#include "vec/vec3.cpp"
-#include "vec/vec4.cpp"
+#include "mat/mat3x3.hpp"
+#include "mat/mat4x4.hpp"
+#include "vec/vec2.hpp"
+#include "vec/vec3.hpp"
+#include "vec/vec4.hpp"
 
 template<typename T>
 static void print(const std::string& name, const vec2<T>& vector) {
@@ -22,21 +22,21 @@ static void print(const std::string& name, const vec4<T>& vector) {
 
 static void test1() {
 	// (OK) Default constructor
-	vec2<int> v1(1, 2);
+	vec2<float> v1(1, 2);
 
 	// (OK) Copy constructor
-	vec2<int> v2 = v1;
+	vec2<float> v2 = v1;
 	// OR
-	//vec2<int> v2(v1);
+	//vec2<float> v2(v1);
 
 	// (OK) Addition
 	//v1 + v2;
 
 	// (OK) Addition, then assignment
 	// This calls the constructor once per addition!
-	//vec2<int> v3(v1 + v2);
+	//vec2<float> v3(v1 + v2);
 	// OR
-	//vec2<int> v3 = v1 + v2;
+	//vec2<float> v3 = v1 + v2;
 
 	// (OK) Addition assignment
 	// This does not call the constructor
@@ -44,9 +44,9 @@ static void test1() {
 
 	// (OK) Addition assignment, then assignment
 	// This calls the constructor once
-	vec2<int> v3 = v1 += v2;
+	vec2<float> v3 = v1 += v2;
 	// OR
-	//vec2<int> v3(v1 += v2);
+	//vec2<float> v3(v1 += v2);
 
 	print("v1", v1);
 	print("v2", v2);
@@ -70,9 +70,9 @@ static void testAdd() {
 }
 
 static void testVec2Sum() {
-	vec2<uint32_t> a(1, 2);
-	vec2<uint32_t> b(3, 4);
-	vec2<uint32_t> c(0, 0);
+	vec2<float> a(1, 2);
+	vec2<float> b(3, 4);
+	vec2<float> c(0, 0);
 
 	print("c", c);
